Actualize develop

+++ chatapi/src/main/java/sk/avo/chatapi/domain/user/IUserRepo.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface IUserRepo extends JpaRepository<UserModel, Integer> {
+public interface IUserRepo extends JpaRepository<UserModel, Long> {
     Optional<UserModel> findByUsername(String username);
     Optional<UserModel> findByEmail(String email);
 }
